WIP => HACK ALL THE THINGS
Starting off with a TDD approach with history.c after I realized that
readline.c and readline.h are full of crazy. Given that I'd like to
implement the minimum viable parts of readline initially and flesh it
out later, it seems like identifying the useful parts of libreadline and
marking everything else as not implemented is probably the easiest way of
achieving something useful in a short amount of time.